chore: Reorganize providers
