+++ src/providers/Providers.tsx
@@ -1,13 +1,11 @@
 'use client';
 
 import ThemeProvider from './ThemeProvider';
-import { GA4TrackerProvider } from '../contexts/ga4Tracker';
-import TagManager from 'react-gtm-module';
 import ApolloContainer from '../components/ApolloContainer';
 import dynamic from 'next/dynamic';
-import { FC, ReactNode, useEffect, useState } from 'react';
+import { FC, ReactNode } from 'react';
 import EnsProvider from './EnsProvider';
-import { AddressEns } from '../services/server/ens/types';
+import TrackerProvider from './TrackerProvider';
 
 const FeatureFlagProvider = dynamic(() => import('../utils/featureFlags'), {
     ssr: false,
@@ -17,50 +15,19 @@ const Web3Container = dynamic(() => import('../components/Web3Container'), {
     ssr: false,
 });
 
-const getENSCachedData = async () => {
-    const host = location.origin;
-    const endpoint = `${host}/api/mainnet/ens`;
-    try {
-        const response = await fetch(endpoint, {
-            signal: AbortSignal.timeout(3000),
-        });
-        return await response.json();
-    } catch (reason) {
-        console.error(
-            `Fetching ENS cached data failed.\nReason: ${reason.message}`
-        );
-        return { data: [] };
-    }
-};
-
 interface ProvidersProps {
     children: ReactNode;
 }
 
 const Providers: FC<ProvidersProps> = ({ children }) => {
-    const [ensData, setEnsData] = useState<AddressEns[]>([]);
-
-    useEffect(() => {
-        if (process.env.NODE_ENV === 'production') {
-            TagManager.initialize({
-                gtmId: process.env.NEXT_PUBLIC_GOOGLE_TAG_MANAGER_ID,
-            });
-        }
-
-        (async () => {
-            const { data }: { data: AddressEns[] } = await getENSCachedData();
-            setEnsData(data);
-        })();
-    }, []);
-
     return (
         <ThemeProvider>
             <FeatureFlagProvider>
-                <EnsProvider ensData={ensData}>
+                <EnsProvider>
                     <Web3Container>
-                        <GA4TrackerProvider>
+                        <TrackerProvider>
                             <ApolloContainer>{children}</ApolloContainer>
-                        </GA4TrackerProvider>
+                        </TrackerProvider>
                     </Web3Container>
                 </EnsProvider>
             </FeatureFlagProvider>
